backend/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
Naked==0.1.32
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
py-diffie-hellman==1.0.1
pycryptodome==3.20.0
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-barcode==0.15.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
shiboken6==6.7.2
six==1.16.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
